+++ trunk/linux-2.6.x/arch/armnommu/mach-nds/arm7/spi.h
@@ -2,5 +2,5 @@
 #define SPI_H
 u16 touchRead(u32 command);
 void poweroff( void );
-void read_firmware( u32 address, u16 * destination, int count);
+void read_firmware( u32 address, u8 * destination, int count);
 #endif
